template: move App to popup dir
