+++ packages/react-scripts/template/src/popup/App.css
