+++ packages/react-scripts/template/src/popup/App.js
