+++ packages/react-scripts/template/src/popup/App.test.js
